splitting up heatmap file - untested
